# feat(search_family): Introduce SingleBlockRangeResult and TwoBlocksRangeResult