… src/core/search/range_tree_test.cc
… src/core/search/search_test.cc