TODO: PR description


Before:

| Benchmark                                                   | Time (ns) | CPU (ns) | Iterations |
|-------------------------------------------------------------|-----------|----------|------------|
| BM_SearchSeveralNumericAndTagIndexes/num_docs:10000         |   345431  |  345381  |     2021    |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:100000        |  3132211  | 3132008  |      224    |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:1000000       | 45877866  |45861242  |       15    |


After:

| Benchmark                                                   | Time (ns) | CPU (ns) | Iterations |
|-------------------------------------------------------------|-----------|----------|------------|
| BM_SearchSeveralNumericAndTagIndexes/num_docs:10000         |   208139  |  208107  |     3352    |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:100000        |  2642775  | 2642433  |      263    |
| BM_SearchSeveralNumericAndTagIndexes/num_docs:1000000       | 42896361  |42891550  |       17    |